[tokenizer] Add support to convert huggingface model to onnx model (#3093)
* [tokenizer] Add support to convert huggingface model to onnx model

* Fix metadata for onnx

* Uses onnx import cli

---------

+++ extensions/tokenizers/src/main/python/model_zoo_importer.py
@@ -11,6 +11,7 @@
 # BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, express or implied. See the License for
 # the specific language governing permissions and limitations under the License.
 import logging
+import os.path
 import shutil
 import sys
 
@@ -63,7 +64,8 @@ def main():
 
         huggingface_models.update_progress(model_info, converter.application,
                                            result, reason, size, args.cpu_only)
-        shutil.rmtree(temp_dir)
+        if os.path.exists(temp_dir):
+            shutil.rmtree(temp_dir)
 
     logging.info("finished.")
